moved AppointmentDetail to components/appointments
